Пытался написать тесты

+++ repository/src/main/java/by/it/academy/task10/dao/Dao.java
@@ -56,11 +56,12 @@ public class Dao<T> implements GenericDao<T> {
     }
 
     @Override
-    public void deleteById(final long entityId) throws SQLException {
+    public void deleteById(final Integer entityId) throws SQLException {
         final T entity = findOne(entityId);
         entityManager.getTransaction().begin();
         delete(entity);
         entityManager.getTransaction().commit();
+
     }
 
     @Override
